requirements.txt
py-cord
yt-dlp
python-dotenv